+++ database/README.md
@@ -1,11 +1,13 @@
 # **Relational Database System:**
 
+**SQLite** View Standalone with [SQLite Studio](https://sqlitestudio.pl/index.rvt)
+
 ## **Consists out of 3 Tables:**
 ### **Clients Table**
 | userId  | Name  | Surname  | E-mail  | Password  |
 |---|---|---|---|---|
 
-### **Audit Table**
+### **Audit Table** WIP
 | userId  | actionPerformed  | timeAccessed  | 
 |---|---|---|
