+++ package.json
@@ -1,20 +1,17 @@
 {
   "name": "group7server",
-  "version": "0.0.1",
-  "description": "Merlot client information system",
-  "main": "index.js",
+  "version": "0.0.0",
+  "description": "backend for merlot client information system",
+  "main": "app.js",
   "dependencies": {
     "express": "^4.16.4",
-    "mock-express-response": "^0.2.2",
-    "pg": "^7.9.0"
+    "sqlite3": "^4.0.6"
   },
-  "devDependencies": {
-    "jest": "^24.5.0"
-  },
+  "devDependencies": {"jest": "^24.5.0"},
   "scripts": {
     "test": "jest",
-    "start": "node index.js"
+    "start": "node app.js"
   },
-  "author": "COS301 - Group 7 Merlot",
+  "author": "COS301 - group 7 merlot",
   "license": "ISC"
 }
