+++ terminal.js
